Getting Started

Normaliz.jl is a Julia interface to Normaliz, an open source tool for computations in affine monoids, vector configurations, lattice polytopes, and rational cones.

Installation

To use Normaliz.jl we require Julia 1.10 or higher. Please see https://julialang.org/downloads/ for instructions on how to obtain julia for your system.

To install this package, enter this into the Julia prompt:

using Pkg; Pkg.develop(url="https://github.com/Normaliz/Normaliz.jl")

Here is an example of using Normaliz.jl:

julia> using Normaliz

julia> C = Cone(; cone = [1 2; 3 5], grading = [1 1])
Normaliz cone

julia> cone_property(C, :HilbertBasis)
2×2 Matrix{BigInt}:
 1  2
 3  5

julia> cone_property(C, :EmbeddingDim)
2

Public API

Normaliz.Cone — Type

Cone(input::AbstractDict; type::Symbol = :gmp)
Cone(; type::Symbol = :gmp, kwargs...)

Construct a Normaliz cone from Julia matrices.

Inputs use Normaliz input names as keyword arguments or as Symbol keys in a dictionary, for example :cone and :grading. Matrix entries must be convertible to rational numbers. By default, Normaliz uses arbitrary precision integer arithmetic; pass type = :longlong to use 64-bit integer arithmetic.

Examples

julia> using Normaliz

julia> C = Cone(; cone = [1 2; 3 5], grading = [1 1])
Normaliz cone

julia> cone_property(C, :HilbertBasis)
2×2 Matrix{BigInt}:
 1  2
 3  5

Normaliz.cone_property — Function

cone_property(cone::Cone, property::Union{AbstractString,Symbol})

Compute and return property for cone.

The result is converted to ordinary Julia objects where this is supported, for example Matrix{BigInt}, Vector{BigInt}, BigInt, Rational{BigInt}, Float64, Int, or Bool.

Examples

julia> using Normaliz

julia> C = Cone(; cone = [1 2; 3 5], grading = [1 1]);

julia> cone_property(C, :HilbertBasis)
2×2 Matrix{BigInt}:
 1  2
 3  5

julia> cone_property(C, :EmbeddingDim)
2

Normaliz.known_cone_properties — Function

known_cone_properties() -> Vector{String}

Return the names of cone properties known to Normaliz.

Examples

julia> using Normaliz

julia> "HilbertBasis" in known_cone_properties()
true

Normaliz.computed_cone_properties — Function

computed_cone_properties(cone::Cone) -> Vector{String}

Return the names of properties that have already been computed for cone.

Examples

julia> using Normaliz

julia> C = Cone(; cone = [1 2; 3 5]);

julia> "ExtremeRays" in computed_cone_properties(C)
false

julia> cone_property(C, :ExtremeRays);

julia> "ExtremeRays" in computed_cone_properties(C)
true

Normaliz.is_computed — Function

is_computed(cone::Cone, property::Union{AbstractString,Symbol}) -> Bool

Return whether property has already been computed for cone.

Examples

julia> using Normaliz

julia> C = Cone(; cone = [1 2; 3 5]);

julia> is_computed(C, :ExtremeRays)
false

julia> cone_property(C, :ExtremeRays);

julia> is_computed(C, :ExtremeRays)
true